Split descriptions.ts into tools/search + tools/execute; add constants.ts
Each codemode tool now owns its registration and description-building, like
tools/docs-search and tools/non-codemode:
- tools/search.ts: registerSearchTool (fetches products, builds the search
  description, registers the tool).
- tools/execute.ts: registerExecuteTool (builds the per-session Cloudflare
  types + account_id param description, registers both account-token and
  user-token shapes).

Truly static, dependency-free values live in constants.ts: SERVER_INFO (+
ServerInfo type, formerly the misnamed ClientInfo) and the CLOUDFLARE_TYPES
declaration block. Replaces the standalone server-info.ts and the grab-bag
descriptions.ts.

server.ts is now ~36 lines of pure wiring: new server -> attach metrics ->
register docs, then either non-codemode or search+execute. No behavior
change (246 tests green).
